added add methods in controller and main class

+++ src/main/kotlin/models/proteinMilk.kt
@@ -1,4 +1,4 @@
 package models
 
-data class proteinMilk(var milkTitle: String, var price:Float, var proteinAmount:Float, var brand:String, var litres:Float) {
+data class proteinMilk(var milkTitle: String, var price:Double, var proteinAmount:Double, var brand:String, var litres:Double) {
 }
